medi_llm/requirements.txt
chromadb==0.5.7
sentence-transformers==3.3.1
torch==2.3.1
anthropic==0.34.2
python-dotenv==1.2.1
gradio==6.6.0
PyYAML==6.0.3
# posthog==3.5.0
